Refactor dev graph timeline views
