+++ tools/dev-graph-ui/src/app/dev-graph/timeline/svg/page.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import { TimelineExperience } from '../TimelineExperience';
+
+export default function SvgTimelinePage() {
+  return (
+    <TimelineExperience
+      initialEngine="svg"
+      allowEngineSwitch={false}
+      alternateEngineHref="/dev-graph/timeline/webgl"
+      alternateEngineLabel="Try the WebGL2 renderer"
+    />
+  );
+}
