+++ tools/dev-graph-ui/src/app/dev-graph/timeline/webgl/page.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import { TimelineExperience } from '../TimelineExperience';
+
+export default function WebglTimelinePage() {
+  return (
+    <TimelineExperience
+      initialEngine="webgl"
+      allowEngineSwitch={false}
+      alternateEngineHref="/dev-graph/timeline/svg"
+      alternateEngineLabel="Switch to the SVG renderer"
+    />
+  );
+}
